requirements.txt
flask
tensorflow-cpu
numpy
joblib
flask-cors
gunicorn